Now this is really starting to look like some professional code.
Parser now took the place of comment parser, and now reads files in and doesn't require code to be passed first.

Plugin now correctly interfaces both the assets and the document plugins. Will call the appropriate method depending on if we're rendering or building.

HTML and Eco document plugins updated to work with the new changes